feat: extract the plugin ABI contract into a tellur-plugin crate
