+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["tellur-core", "tellur-live", "tellur-macros", "tellur-renderer"]
+members = ["tellur-core", "tellur-live", "tellur-macros", "tellur-plugin", "tellur-renderer"]
 resolver = "2"
