+++ tellur-live/Cargo.toml
@@ -6,6 +6,7 @@ build = "build.rs"
 
 [dependencies]
 tellur-core = { path = "../tellur-core" }
+tellur-plugin = { path = "../tellur-plugin" }
 tellur-renderer = { path = "../tellur-renderer" }
 
 [[example]]
