+++ tellur-plugin/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "tellur-plugin"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+tellur-core = { path = "../tellur-core" }
